refactor(gateway): extract interrupt helpers from run.py (slice 19 of #54962)
